+++ packages/cli/src/commands/index.ts
@@ -3,5 +3,7 @@ import genkeys from "./genkeys.js"
 import deploy from "./deploy.js"
 import initialize from "./initialize.js"
 import addRecipients from "./addRecipients.js"
+import signup from "./signup.js"
+import doTheThing from "./doTheThing.js"
 
-export { auth, genkeys, deploy, initialize, addRecipients }
+export { auth, genkeys, deploy, initialize, addRecipients, signup, doTheThing }