+++ packages/cli/src/index.ts
@@ -2,7 +2,7 @@
 import dotenv from "dotenv"
 import { createCommand } from "commander"
 import { readLocalJsonFile } from "./lib/files.js"
-import { auth, genkeys, deploy, initialize, addRecipients } from "./commands/index.js"
+import { auth, genkeys, deploy, initialize, addRecipients, signup, doTheThing } from "./commands/index.js"
 
 dotenv.config()
 
@@ -59,6 +59,23 @@ program
   .argument("<path>", "the path of the CSV input file where the recipients data is stored")
   .action((network: string, path: string) => {
     addRecipients(network, path)
+  })
+
+program
+  .command("contracts:signup")
+  .description("Signup users and refill with some crypto their addresses")
+  .argument("<network>", "the network where the contracts has been deployed")
+  .argument("<path>", "the path of the CSV input file where the signup data for users is stored")
+  .action((network: string, path: string) => {
+    signup(network, path)
+  })
+
+  program
+  .command("dothething")
+  .description("Does all the things without user input")
+  .argument("<network>", "the network where the contracts has been deployed")
+  .action((network: string) => {
+    doTheThing(network)
   })
 
 program.parseAsync()